refactor(hero validation): move RSR validation into that file
